Address some changes
